+++ views/ocho-dias.view.js
@@ -1,0 +1,5 @@
+import { renderVistaAgenda } from '../assets/js/vista-agenda.js';
+
+export function renderVistaOchoDias(contenedor) {
+  renderVistaAgenda(contenedor, 8);
+}