+++ views/tres-dias.view.js
@@ -1,0 +1,5 @@
+import { renderVistaAgenda } from '../assets/js/vista-agenda.js';
+
+export function renderVistaTresDias(contenedor) {
+  renderVistaAgenda(contenedor, 3);
+}